content-security-policy/to-port-from-blink/inline-style-blocked-expected.txt

CONSOLE ERROR: line 5: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-icXySbIcypyRXkmnfqkPP-2rK07ZHUNZ9oTzt6HzoB4='), or a nonce ('nonce-...') is required to enable inline execution.

PASS
